BREAKING CHANGE: cleaned up the nulls package to be a bit more idiomatic.

+++ defaults/defaults_test.go
@@ -28,26 +28,26 @@ func Test_Int(t *testing.T) {
 func Test_Int64(t *testing.T) {
 	a := assert.New(t)
 
-	a.Equal(defaults.Int64(0, 1), 1)
-	a.Equal(defaults.Int64(2, 1), 2)
+	a.Equal(defaults.Int64(0, 1), int64(1))
+	a.Equal(defaults.Int64(2, 1), int64(2))
 	var s int64
-	a.Equal(defaults.Int64(s, 1), 1)
+	a.Equal(defaults.Int64(s, 1), int64(1))
 }
 
 func Test_Float32(t *testing.T) {
 	a := assert.New(t)
 
-	a.Equal(defaults.Float32(0, 1), 1)
-	a.Equal(defaults.Float32(2, 1), 2)
+	a.Equal(defaults.Float32(0, 1), float32(1))
+	a.Equal(defaults.Float32(2, 1), float32(2))
 	var s float32
-	a.Equal(defaults.Float32(s, 1), 1)
+	a.Equal(defaults.Float32(s, 1), float32(1))
 }
 
 func Test_Float64(t *testing.T) {
 	a := assert.New(t)
 
-	a.Equal(defaults.Float64(0, 1), 1)
-	a.Equal(defaults.Float64(2, 1), 2)
+	a.Equal(defaults.Float64(0, 1), float64(1))
+	a.Equal(defaults.Float64(2, 1), float64(2))
 	var s float64
-	a.Equal(defaults.Float64(s, 1), 1)
+	a.Equal(defaults.Float64(s, 1), float64(1))
 }
